week end update

+++ includes/computor.h
@@ -6,7 +6,7 @@
 /*   By: kwiessle <kwiessle@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/23 18:39:06 by kwiessle          #+#    #+#             */
-/*   Updated: 2017/12/04 11:13:49 by vquesnel         ###   ########.fr       */
+/*   Updated: 2017/12/04 11:14:02 by vquesnel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,7 +38,7 @@ typedef struct      s_env
 t_env     *new_env(double a, double b, double c);
 int       key_events(int keycode, t_env *env);
 void      init_graph(t_env *env);
-int       polynomial(int x, double a, double b, double c);
+long long       polynomial(int x, double a, double b, double c);
 void      draw_curve(t_env *env);
 
 /* ERROR */
